+++ diagnostics/tropical_rainfall/tropical_rainfall/src/tropical_rainfall_tools.py
@@ -59,19 +59,6 @@ class ToolsClass:
             time_str = '-'.join(parts[:len(parts)])
         return time_str
 
-    def get_machine(self):
-        """
-        Retrieves the machine information from the ConfigPath instance.
-
-        Returns:
-            str: The machine information retrieved from the ConfigPath object.
-
-        Raises:
-            SomeException: If the ConfigPath object is not properly initialized or if there is an issue with
-                        retrieving the machine information.
-        """
-        return ConfigPath().catalog
-
     def get_netcdf_path(self, configname: str = full_path_to_config) -> tuple:
         """
         Load paths from a YAML configuration file based on the specified configuration name.
@@ -94,7 +81,7 @@ class ToolsClass:
         try:
             with open(config_path, 'r') as file:
                 data = yaml.safe_load(file)
-            machine = ConfigPath().machine
+            machine = ConfigPath().get_machine()
             path_to_netcdf = data[machine]['path_to_netcdf']
         except FileNotFoundError as e:
             # Handle FileNotFoundError exception
@@ -187,7 +174,7 @@ class ToolsClass:
         try:
             with open(config_path, 'r') as file:
                 data = yaml.safe_load(file)
-            machine = ConfigPath().machine
+            machine = ConfigPath().get_machine()
             path_to_pdf = data[machine]['path_to_pdf']
         except FileNotFoundError as e:
             # Handle FileNotFoundError exception
@@ -259,25 +246,38 @@ class ToolsClass:
         except (KeyError, TypeError):
             return default
 
-    def open_dataset(self, path_to_netcdf: str) -> object:
+    def open_dataset(self, path_to_netcdf: str) -> xr.Dataset:
         """
-        Function to load a histogram dataset from a file using pickle.
+        Function to load a dataset from a NetCDF file.
 
         Args:
             path_to_netcdf (str): The path to the dataset file.
 
         Returns:
-            object: The loaded histogram dataset.
+            xr.Dataset: The loaded dataset.
 
         Raises:
             FileNotFoundError: If the specified dataset file is not found.
+            ValueError: If the dataset file cannot be opened with the available backends.
         """
+        self.logger.debug(f"Opening dataset from path: {path_to_netcdf}")
+        
+        if not os.path.exists(path_to_netcdf):
+            self.logger.error(f"File does not exist: {path_to_netcdf}")
+            raise FileNotFoundError(f"File does not exist: {path_to_netcdf}")
+        
         try:
-            dataset = xr.open_dataset(path_to_netcdf)
+            dataset = xr.open_dataset(path_to_netcdf, engine='netcdf4')
             return dataset
-        except FileNotFoundError:
-            raise FileNotFoundError(
-                "The specified dataset file was not found.")
+        except FileNotFoundError as e:
+            self.logger.error(f"File not found: {path_to_netcdf}")
+            raise FileNotFoundError(f"The specified dataset file was not found: {path_to_netcdf}")
+        except ValueError as e:
+            self.logger.error(f"Error opening dataset with available backends: {e}")
+            raise ValueError(f"Failed to open dataset at {path_to_netcdf}: {e}")
+        except Exception as e:
+            self.logger.error(f"Unexpected error opening dataset: {e}")
+            raise
 
     def select_files_by_year_and_month_range(self, path_to_histograms: str, start_year: int = None, end_year: int = None,
                                              start_month: int = None, end_month: int = None, flag: str = None) -> list:
@@ -796,25 +796,29 @@ class ToolsClass:
         filenames = [os.path.basename(file) for file in filenames]
         times = [self.parse_filename_to_datetime(filename) for filename in filenames]
         times.sort(key=lambda x: x[0])  # Sort by start time
-        
+
         for i in range(len(times) - 1):
             current_start_time, current_end_time = times[i]
             next_start_time, _ = times[i + 1]
-            
+
             if freq == 'M':  # Monthly data
                 expected_next_start = current_end_time + relativedelta(months=+1)
-                # Adjust for the end of the month
-                expected_next_start = expected_next_start.replace(day=1, hour=0)
+                expected_next_start = expected_next_start.replace(day=1, hour=0)  # Adjust for the end of the month
             elif '3H' in freq:  # 3-hourly data
                 expected_next_start = current_end_time + relativedelta(hours=+3)
-            
+            else:
+                self.logger.warning(f"Unsupported frequency: {freq}")
+                continue
+
             if next_start_time != expected_next_start:
-                self.logger.error(f"Discontinuity in results found: Expected {expected_next_start}, got {next_start_time}")
+                self.logger.error(f"Discontinuity found: Expected {expected_next_start}, got {next_start_time}")
                 return False
             else:
-                self.logger.info(f"Continuity of results confirmed: {next_start_time} follows {expected_next_start}")
-            
+                self.logger.debug(f"Continuity confirmed: {next_start_time} follows {expected_next_start}")
+
+        self.logger.info("Continuity of time coordinates confirmed for all files.")
         return True
+
     
     def check_incomplete_months(self, files):
         filenames = [os.path.basename(file) for file in files]
@@ -1057,3 +1061,56 @@ class ToolsClass:
             for i in _names:
                 _size *= data[i].size
         return _size
+
+    def format_time(self, time_band: str) -> str:
+        """
+        Extract the dates from the time_band string and format them as 'year-month-day'.
+
+        Args:
+            time_band (str): The input time_band string.
+
+        Returns:
+            str: The formatted date string.
+        """
+        # Find all datetime parts and format them
+        formatted_time_band = re.sub(r'(\d{4})-(\d{2})-(\d{2})T\d{2}:\d{2}:\d{2}\.\d+', r'\1-\2-\3', time_band)
+        return formatted_time_band
+    
+    def format_lat_band(self, dataset) -> str:
+        """
+        Format the latitude band information from a dataset.
+
+        Args:
+            dataset: The dataset containing the latitude band information.
+
+        Returns:
+            str: A formatted string describing the latitude band.
+        """
+        lat_band_values = dataset.lat_band.split(', ')
+        lat_band_str = f"The latitude band is from {lat_band_values[0]} to {lat_band_values[1]}"
+        if len(lat_band_values) > 2:
+            lat_band_str += f" with a frequency of {lat_band_values[2].split('=')[1]}"
+        return lat_band_str
+
+    def verify_lat_band(sel, model_dataset, comparison_dataset, dataset_name, logger):
+        """
+        Verify that the latitude band of a comparison dataset matches the latitude band of the model dataset.
+
+        Args:
+            model_dataset: The model dataset containing the reference latitude band.
+            comparison_dataset: The dataset to be compared.
+            dataset_name: The name of the comparison dataset.
+            logger: The logger instance for logging messages.
+
+        Returns:
+            bool: True if the latitude bands match, False otherwise.
+        """
+        def check_lat_band(model_lat_band, comparison_lat_band, dataset_name):
+            if model_lat_band != comparison_lat_band:
+                logger.error(f"Latitude bands do not match between model and {dataset_name} data. Aborting comparison.")
+                return False
+            return True
+
+        if comparison_dataset is not None:
+            return check_lat_band(model_dataset.lat_band, comparison_dataset.lat_band, dataset_name)
+        return True  # If comparison_dataset is None, consider it as a match